+++ test/test_files/test_error_expected/selected_names/test_wrong_name.vhd
@@ -1,0 +1,9 @@
+entity test_wrong_name is
+end entity;
+architecture arch of test_wrong_name is
+  signal x : integer;
+
+begin
+  x <= x.all;
+
+end architecture;
